+++ packages/service-fastify-static/src/hooks.js
@@ -1,0 +1,5 @@
+const SERVICE_NAME = `fastify-static`;
+
+module.exports = {
+  SERVICE_NAME,
+};