+++ packages/service-fastify-static/src/lib/hooks.js
@@ -1,7 +1,0 @@
-const { SERVICE } = require('@forrestjs/hooks');
-
-const SERVICE_NAME = `${SERVICE} fastify-static`;
-
-module.exports = {
-  SERVICE_NAME,
-};