+++ CHANGELOG.md
@@ -35,9 +35,12 @@
   - Add extendable `/test` route for E2E test support
   - Added globals utilities for E2E testing in Jest
   - Integrates with `service-env` to provide direct access to `getEnv` in the route handlers
+- service-fastify-static
+  - Remove Babel
 - kitchensink
   - Added examples for `@forrestjs/hooks`
   - Added examples for `@forrestjs/service-fastify`
+  - Added examples for `@forrestjs/service-fastify-static`
   - Added examples for `@forrestjs/service-jwt`
   - Added examples for `@forrestjs/service-hash`
   - Added examples for `@forrestjs/service-fetchq`